refactor(Tests Extraction): 正交拆分 test_extraction.py（1845行）为 contracts/llm_plan/provider 三个单责模块;
- test_extraction_contracts.py: resolve_source_kind, resolve_targets, build_tool_adapter 等契约函数（7 tests）
- test_extraction_llm_plan.py: _build_llm_invocation_plan 全部场景（5 tests）
- test_extraction_provider.py: DataExtractorProvider._invoke_target 和 .extract（15 tests）
- 15 个内联 FakeSession 替换为 conftest 的参数化 FakeMcpSession
